64 bit patches from Zach Brown sprintf replaced with snprintf
git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc/trunk@48 315857ad-0bdb-0310-b42e-dec37551a5f0

+++ vpnc.c
@@ -468,8 +468,8 @@ static uint16_t unpack_verify_phase2(struct sa_block *s,
 	*r_p = NULL;
 
 	if (r_length < ISAKMP_PAYLOAD_O || ((r_length - ISAKMP_PAYLOAD_O) % s->ivlen != 0)) {
-		DEBUG(2, printf("payload to short or not padded: len=%d, min=%d (ivlen=%d)\n",
-			r_length, ISAKMP_PAYLOAD_O, s->ivlen));
+		DEBUG(2, printf("payload to short or not padded: len=%lld, min=%d (ivlen=%lld)\n",
+			(long long)r_length, ISAKMP_PAYLOAD_O, (long long)s->ivlen));
 		return ISAKMP_N_UNEQUAL_PAYLOAD_LENGTHS;
 	}
